# apps/api/internal/http/handlers/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Enterprise File Platform API",
    "version": "0.1.0",
    "description": "Local-first ingestion orchestration and search API for file parsing, job execution, and operational monitoring.",
    "contact": {
      "name": "Project Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development API"
    }
  ],
  "paths": {
...
    "/api/v1/uploads/{upload_id}": {
      "get": {
        "summary": "Get upload metadata",
        "operationId": "getUploadByID",
        "tags": ["Uploads"],
        "parameters": [
          {"name": "upload_id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {
            "description": "Upload metadata",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Upload"},
                "example": {
                  "id": "d3ef1ac4-2f4f-4d8d-95d5-a7f3f5a8eb4a",
                  "tenant_id": "11111111-1111-1111-1111-111111111001",
                  "status": "pending",
                  "original_source": "sample.zip",
                  "storage_prefix": "uploads/11111111-1111-1111-1111-111111111001/d3ef1ac4-2f4f-4d8d-95d5-a7f3f5a8eb4a",
                  "total_files": 1,
                  "total_size_bytes": 12345,
                  "metadata": {
                    "file_name": "sample.zip",
                    "size_bytes": 12345,
                    "content_type": "application/zip",
                    "password_provided": false
                  },
                  "created_at": "2026-06-16T11:40:12Z",
                  "updated_at": "2026-06-16T11:40:13Z"
                }
              }
            }
          },
          "404": {"$ref": "#/components/responses/NotFound"},
          "500": {"$ref": "#/components/responses/InternalError"}
        }
      }
    },
...
    "/api/v1/files/{file_id}/tree": {
      "get": {
        "summary": "Get recursive file tree rooted at a file",
        "operationId": "getFileTree",
        "tags": ["Files"],
        "parameters": [
          {"name": "file_id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {
            "description": "Nested file tree",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/FileTreeNode"},
                "example": {
                  "file": {
                    "id": "f6beef3c-8a45-4b89-8a3a-3d6d5b2d7f88",
                    "tenant_id": "11111111-1111-1111-1111-111111111001",
                    "original_name": "sample.zip",
                    "size_bytes": 12345,
                    "extension": ".zip",
                    "detected_file_type": "archive",
                    "storage_path": "uploads/11111111-1111-1111-1111-111111111001/d3ef1ac4-2f4f-4d8d-95d5-a7f3f5a8eb4a/sample.zip",
                    "depth": 0,
                    "is_archive": true,
                    "processing_status": "completed",
                    "created_at": "2026-06-16T11:40:12Z",
                    "updated_at": "2026-06-16T11:41:00Z"
                  },
                  "children": []
                }
              }
            }
          },
          "404": {"$ref": "#/components/responses/NotFound"},
          "500": {"$ref": "#/components/responses/InternalError"}
        }
      }
    },
...
    "/api/v1/jobs": {
      "get": {
        "summary": "List jobs",
        "operationId": "listJobs",
        "tags": ["Jobs"],
        "parameters": [
          {"name": "tenant_id", "in": "query", "required": false, "schema": {"type": "string"}},
          {"name": "page", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 1}, "example": 1},
          {"name": "page_size", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 1, "maximum": 200}, "example": 25}
        ],
        "responses": {
          "200": {
            "description": "Paginated jobs",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/JobListResponse"},
                "example": {
                  "total": 2,
                  "page": 1,
                  "page_size": 25,
                  "jobs": [
                    {
                      "id": "6d5fbf3a-21f8-4f2f-8bf7-cd6e9f13d9d6",
                      "tenant_id": "11111111-1111-1111-1111-111111111001",
                      "root_file_id": "f6beef3c-8a45-4b89-8a3a-3d6d5b2d7f88",
                      "status": "completed",
                      "current_stage": "completed",
                      "progress_percent": 100,
                      "retry_count": 0,
                      "created_at": "2026-06-16T11:40:13Z",
                      "updated_at": "2026-06-16T11:41:20Z"
                    }
                  ]
                }
              }
            }
          },
          "500": {"$ref": "#/components/responses/InternalError"}
        }
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "summary": "Get job",
        "operationId": "getJobByID",
        "tags": ["Jobs"],
        "parameters": [
          {"name": "job_id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Job", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/IngestionJob"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "500": {"$ref": "#/components/responses/InternalError"}
        }
      }
    },
...
    "/api/v1/search": {
      "get": {
        "summary": "Search parsed records",
        "operationId": "searchRecordsGet",
        "tags": ["Search"],
        "parameters": [
          {"name": "q", "in": "query", "required": false, "schema": {"type": "string"}, "description": "Full text query", "example": "authentication"},
          {"name": "file_id", "in": "query", "required": false, "schema": {"type": "string"}},
          {"name": "extension", "in": "query", "required": false, "schema": {"type": "string", "example": "txt"}},
          {"name": "detected_file_type", "in": "query", "required": false, "schema": {"type": "string", "example": "text"}},
          {"name": "record_type", "in": "query", "required": false, "schema": {"type": "string", "example": "text_line"}},
          {"name": "date_from", "in": "query", "required": false, "schema": {"type": "string"}},
          {"name": "date_to", "in": "query", "required": false, "schema": {"type": "string"}},
          {"name": "ip", "in": "query", "required": false, "schema": {"type": "string", "format": "ipv4"}},
          {"name": "email", "in": "query", "required": false, "schema": {"type": "string"}},
          {"name": "domain", "in": "query", "required": false, "schema": {"type": "string"}},
          {"name": "job_id", "in": "query", "required": false, "schema": {"type": "string"}},
          {"name": "tenant_id", "in": "query", "required": false, "schema": {"type": "string"}},
          {"name": "page", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 1}, "example": 1},
          {"name": "page_size", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 1, "maximum": 100}, "example": 25},
          {"name": "sort", "in": "query", "required": false, "schema": {"type": "string", "enum": ["relevance", "created_at"]}}
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/SearchResponse"},
                "example": {
                  "total": 1,
                  "page": 1,
                  "page_size": 25,
                  "results": [
                    {
                      "record_id": "9f5b4f0f-f7d8-4b9f-a1c4-2dd7d5a7d111",
                      "file_id": "f6beef3c-8a45-4b89-8a3a-3d6d5b2d7f88",
                      "job_id": "6d5fbf3a-21f8-4f2f-8bf7-cd6e9f13d9d6",
                      "source_file_name": "sample.txt",
                      "record_type": "text_line",
                      "content_preview": "user=john@example.com ip=198.51.100.12",
                      "highlight": "<em>john</em>@example.com",
                      "entities": {"emails": ["john@example.com"], "ip_addresses": ["198.51.100.12"]},
                      "created_at": "2026-06-16T11:41:20Z"
                    }
                  ],
                  "facets": {
                    "record_type": [{"value": "text_line", "count": 1}],
                    "detected_file_type": [{"value": "txt", "count": 1}],
                    "entities": {
                      "ip_addresses": [{"value": "198.51.100.12", "count": 1}],
                      "emails": [{"value": "john@example.com", "count": 1}],
                      "domains": [{"value": "example.com", "count": 1}],
                      "urls": [],
                      "hashes": []
                    }
                  }
                }
              }
            }
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "500": {"$ref": "#/components/responses/InternalError"}
        }
      },
      "post": {
        "summary": "Search parsed records via request body",
        "operationId": "searchRecordsPost",
        "tags": ["Search"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/SearchRecordsRequest"},
              "example": {
                "q": "john@example.com",
                "email": "john@example.com",
                "page": 1,
                "page_size": 25,
                "sort": "created_at"
              }
            }
          }
        },
        "responses": {
          "200": {"$ref": "#/components/responses/SearchResponseRef"},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "500": {"$ref": "#/components/responses/InternalError"}
        }
      }
    },
...
    "/api/v1/dashboard/summary": {
      "get": {
        "summary": "Get dashboard summary",
        "operationId": "getDashboardSummary",
        "tags": ["Dashboard"],
        "parameters": [{"name": "tenant_id", "in": "query", "required": false, "schema": {"type": "string"}}],
        "responses": {
          "200": {
            "description": "Dashboard summary",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/DashboardSummary"},
                "example": {
                  "tenant_id": "11111111-1111-1111-1111-111111111001",
                  "total_uploads": 3,
                  "total_files": 12,
                  "total_extracted_files": 10,
                  "total_parsed_records": 540,
                  "completed_jobs": 2,
                  "failed_jobs": 0,
                  "password_required_files": 1,
                  "quarantined_files": 0
                }
              }
            }
          },
          "500": {"$ref": "#/components/responses/InternalError"}
        }
      }
    },
...
    "/api/v1/admin/settings": {
      "get": {
        "summary": "Get tenant settings",
        "operationId": "getAdminSettings",
        "tags": ["Admin"],
        "parameters": [{"name": "tenant_id", "in": "query", "required": false, "schema": {"type": "string"}}],
        "responses": {
          "200": {
            "description": "Tenant settings",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/AdminSettings"},
                "example": {
                  "tenant_id": "11111111-1111-1111-1111-111111111001",
                  "max_upload_size_mb": 512,
                  "max_archive_depth": 10,
                  "max_extracted_files": 10000,
                  "max_extracted_size_mb": 1024,
                  "txt_small_file_limit_mb": 10,
                  "max_expansion_ratio": 20,
                  "enabled_parsers": ["txt", "log", "csv", "json", "jsonl"],
                  "parser_batch_size": 1000,
                  "search_index_batch_size": 1000
                }
              }
            }
          },
          "500": {"$ref": "#/components/responses/InternalError"}
        }
      },
      "put": {
        "summary": "Update tenant settings",
        "operationId": "updateAdminSettings",
        "tags": ["Admin"],
        "parameters": [{"name": "tenant_id", "in": "query", "required": false, "schema": {"type": "string"}}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/UpdateAdminSettingsRequest"},
              "example": {"max_upload_size_mb": 1024, "enabled_parsers": ["txt", "csv", "json"]}
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated settings",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/AdminSettings"}
              }
            }
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "403": {
            "description": "Not authorized",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}, "example": {"error": "forbidden: insufficient permissions"}}
            }
          },
          "500": {"$ref": "#/components/responses/InternalError"}
        }
      }
    }
  },
  "components": {
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/Error"},
            "example": {"error": "invalid request"}
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/Error"},
            "example": {"error": "resource not found"}
          }
        }
      },
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/Error"},
            "example": {"error": "internal server error"}
          }
        }
      },
      "SearchResponseRef": {
        "description": "Search results",
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/SearchResponse"}
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {"type": "string"}
        },
        "required": ["error"]
      },
...
      "Upload": {
        "type": "object",
        "properties": {
          "id": {"type": "string"},
          "tenant_id": {"type": "string"},
          "initiated_by": {"type": "string", "nullable": true},
          "status": {"type": "string"},
          "original_source": {"type": "string", "nullable": true},
          "storage_prefix": {"type": "string", "nullable": true},
          "total_files": {"type": "integer"},
          "total_size_bytes": {"type": "integer", "format": "int64"},
          "metadata": {"type": "object", "additionalProperties": true},
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      },
      "File": {
        "type": "object",
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "tenant_id": {"type": "string"},
          "parent_file_id": {"type": "string", "format": "uuid", "nullable": true},
          "upload_id": {"type": "string", "format": "uuid", "nullable": true},
          "original_name": {"type": "string"},
          "normalized_name": {"type": "string", "nullable": true},
          "extension": {"type": "string", "nullable": true},
          "detected_mime_type": {"type": "string", "nullable": true},
          "detected_file_type": {"type": "string", "nullable": true},
          "storage_path": {"type": "string"},
          "size_bytes": {"type": "integer", "format": "int64"},
          "sha256_hash": {"type": "string", "nullable": true},
          "depth": {"type": "integer"},
          "is_archive": {"type": "boolean"},
          "is_password_protected": {"type": "boolean"},
          "processing_status": {"type": "string"},
          "created_by": {"type": "string", "nullable": true},
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      },
...
      "FileTreeNode": {
        "type": "object",
        "properties": {
          "file": {"$ref": "#/components/schemas/File"},
          "children": {"type": "array", "items": {"$ref": "#/components/schemas/FileTreeNode"}}
        }
      },
...
      "IngestionJob": {
        "type": "object",
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "tenant_id": {"type": "string"},
          "root_file_id": {"type": "string", "format": "uuid"},
          "status": {"type": "string"},
          "current_stage": {"type": "string", "nullable": true},
          "progress_percent": {"type": "number", "format": "float"},
          "retry_count": {"type": "integer"},
          "error_code": {"type": "string", "nullable": true},
          "error_message": {"type": "string", "nullable": true},
          "started_at": {"type": "string", "format": "date-time", "nullable": true},
          "completed_at": {"type": "string", "format": "date-time", "nullable": true},
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      },
      "JobListResponse": {
        "type": "object",
        "properties": {
          "total": {"type": "integer", "format": "int64"},
          "page": {"type": "integer"},
          "page_size": {"type": "integer"},
          "jobs": {"type": "array", "items": {"$ref": "#/components/schemas/IngestionJob"}}
        }
      },
...
      "SearchRecordsRequest": {
        "type": "object",
        "properties": {
          "q": {"type": "string"},
          "file_id": {"type": "string"},
          "extension": {"type": "string"},
          "detected_file_type": {"type": "string"},
          "record_type": {"type": "string"},
          "date_from": {"type": "string", "format": "date-time"},
          "date_to": {"type": "string", "format": "date-time"},
          "ip": {"type": "string"},
          "email": {"type": "string"},
          "domain": {"type": "string"},
          "job_id": {"type": "string"},
          "page": {"type": "integer", "minimum": 1},
          "page_size": {"type": "integer", "minimum": 1, "maximum": 100},
          "sort": {"type": "string", "enum": ["relevance", "created_at"]},
          "tenant_id": {"type": "string"}
        }
      },
      "SearchFacetItem": {
        "type": "object",
        "properties": {
          "value": {"type": "string"},
          "count": {"type": "integer", "format": "int64"}
        }
      },
      "SearchResult": {
        "type": "object",
        "properties": {
          "record_id": {"type": "string", "format": "uuid"},
          "file_id": {"type": "string", "format": "uuid"},
          "job_id": {"type": "string", "format": "uuid"},
          "source_file_name": {"type": "string"},
          "record_type": {"type": "string"},
          "content_preview": {"type": "string"},
          "highlight": {"type": "string"},
          "entities": {"type": "object", "additionalProperties": true},
          "created_at": {"type": "string", "format": "date-time"}
        }
      },
      "SearchFacets": {
        "type": "object",
        "properties": {
          "record_type": {"type": "array", "items": {"$ref": "#/components/schemas/SearchFacetItem"}},
          "detected_file_type": {"type": "array", "items": {"$ref": "#/components/schemas/SearchFacetItem"}},
          "entities": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {"$ref": "#/components/schemas/SearchFacetItem"}
            }
          }
        }
      },
      "SearchResponse": {
        "type": "object",
        "properties": {
          "total": {"type": "integer", "format": "int64"},
          "page": {"type": "integer"},
          "page_size": {"type": "integer"},
          "results": {"type": "array", "items": {"$ref": "#/components/schemas/SearchResult"}},
          "facets": {"$ref": "#/components/schemas/SearchFacets"}
        }
      },
...
      "DashboardSummary": {
        "type": "object",
        "properties": {
          "tenant_id": {"type": "string"},
          "total_uploads": {"type": "integer", "format": "int64"},
          "total_files": {"type": "integer", "format": "int64"},
          "total_extracted_files": {"type": "integer", "format": "int64"},
          "total_parsed_records": {"type": "integer", "format": "int64"},
          "completed_jobs": {"type": "integer", "format": "int64"},
          "failed_jobs": {"type": "integer", "format": "int64"},
          "password_required_files": {"type": "integer", "format": "int64"},
          "quarantined_files": {"type": "integer", "format": "int64"}
        }
      },
...
      "AdminSettings": {
        "type": "object",
        "properties": {
          "tenant_id": {"type": "string"},
          "max_upload_size_mb": {"type": "number"},
          "max_archive_depth": {"type": "integer", "format": "int64"},
          "max_extracted_files": {"type": "integer", "format": "int64"},
          "max_extracted_size_mb": {"type": "integer", "format": "int64"},
          "txt_small_file_limit_mb": {"type": "number"},
          "max_expansion_ratio": {"type": "number"},
          "enabled_parsers": {"type": "array", "items": {"type": "string"}},
          "parser_batch_size": {"type": "integer", "format": "int64"},
          "search_index_batch_size": {"type": "integer", "format": "int64"}
        }
      },
      "UpdateAdminSettingsRequest": {
        "type": "object",
        "properties": {
          "max_upload_size_mb": {"type": "number"},
          "max_archive_depth": {"type": "integer", "format": "int64"},
          "max_extracted_files": {"type": "integer", "format": "int64"},
          "max_extracted_size_mb": {"type": "integer", "format": "int64"},
          "txt_small_file_limit_mb": {"type": "number"},
          "max_expansion_ratio": {"type": "number"},
          "enabled_parsers": {"type": "array", "items": {"type": "string"}},
          "parser_batch_size": {"type": "integer", "format": "int64"},
          "search_index_batch_size": {"type": "integer", "format": "int64"}
        }
      }
    }
  },
...
}